Moved polar plot files to own sub-directory so iframe_figures doesn't conflict with other future notebooks.
Each notebook will have it's own iframe_figures sub-directory with it's own figure files but these are named after cell numbers so if there are multiple notebooks, these files could clash and overwrite each other.

The easiest solution seems to be to move the notebook and all it's associated html files into it's own directory. The directory is given the same name as the original html notebook (without the .html extension) and the main html file is then moved to this sub-directory and renamed to index.html. Then links to the main html file won't need to change (as long as it doesn't include the .html extension which is the preferred way to do url's these days).

Links to the notebook itself will need to change to include the sub-directory in the path. I could have left the notebook file where it was and not needed to update any links, but I prefer to keep all the files together in their own sub-directory. This is not much of a problem though since there aren't many links out there pointing to the notebook anyway. I can just update them manually.